Export RemoveRootAttributes method to allow user mode to remove root attributes
When the src directory is moved to the backup location during a dehydrate
the files in the backup folder are empty placeholders cannot be deleted
because the kext thinks it is an offline root and no IO should happen.

Need this method with the dehydrate to remove the attributes from the src
directory in the backup location on Mac so that all files that get moved
to the backup will be able to be deleted/moved.